関数の循環参照エラー、関数の切り出しと適用
